#178 [refactor] domain 모듈로 repository 이동 및 domain model로 리팩토링
